+++ src/main/java/com/malina_ink/resaleplatform/WebSecurityConfig.java
@@ -1,6 +1,4 @@
-package ru.skypro.homework;
-
-import static org.springframework.security.config.Customizer.withDefaults;
+package com.malina_ink.resaleplatform;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,6 +9,8 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
+
+import static org.springframework.security.config.Customizer.withDefaults;
 
 @Configuration
 public class WebSecurityConfig {
