+++ src/main/java/com/malina_ink/resaleplatform/controller/AuthController.java
@@ -1,4 +1,44 @@
 package com.malina_ink.resaleplatform.controller;
 
+import com.malina_ink.resaleplatform.dto.LoginDto;
+import com.malina_ink.resaleplatform.dto.RegisterDto;
+import com.malina_ink.resaleplatform.enums.Role;
+import com.malina_ink.resaleplatform.service.AuthService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import static com.malina_ink.resaleplatform.enums.Role.USER;
+
+@Slf4j
+@CrossOrigin(value = "http://localhost:3000")
+@RestController
+@RequiredArgsConstructor
 public class AuthController {
+
+    private final AuthService authService;
+
+    @PostMapping("/login")
+    public ResponseEntity<?> login(@RequestBody LoginDto req) {
+        if (authService.login(req.getUsername(), req.getPassword())) {
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
+
+    @PostMapping("/register")
+    public ResponseEntity<?> register(@RequestBody RegisterDto req) {
+        Role role = req.getRole() == null ? USER : req.getRole();
+        if (authService.register(req, role)) {
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+    }
+}
