+++ src/main/java/com/malina_ink/resaleplatform/controller/BasicAuthCorsFilter.java
@@ -1,4 +1,24 @@
 package com.malina_ink.resaleplatform.controller;
 
-public class BasicAuthCorsFilter {
+
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Component
+public class BasicAuthCorsFilter extends OncePerRequestFilter {
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest httpServletRequest,
+                                    HttpServletResponse httpServletResponse,
+                                    FilterChain filterChain)
+            throws ServletException, IOException {
+        httpServletResponse.addHeader("Access-Control-Allow-Credentials", "true");
+        filterChain.doFilter(httpServletRequest, httpServletResponse);
+    }
 }
