+++ src/main/java/com/malina_ink/resaleplatform/controller/ImageController.java
@@ -1,0 +1,18 @@
+package com.malina_ink.resaleplatform.controller;
+
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+@RestController
+@RequestMapping("/image")
+public class ImageController {
+//    @PutMapping("/{id}")
+//    public String update(@RequestBody MultipartFile image) {
+////    вызов метода
+//        Image image1 = imageService.getId().updateImage(image);
+//
+//        return String pathToPhoto;
+}
