+++ src/main/java/com/malina_ink/resaleplatform/service/impl/AuthServiceImpl.java
@@ -1,13 +1,13 @@
-package ru.skypro.homework.service.impl;
+package com.malina_ink.resaleplatform.service.impl;
 
+import com.malina_ink.resaleplatform.dto.RegisterDto;
+import com.malina_ink.resaleplatform.enums.Role;
+import com.malina_ink.resaleplatform.service.AuthService;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
-import ru.skypro.homework.dto.RegisterReq;
-import ru.skypro.homework.dto.Role;
-import ru.skypro.homework.service.AuthService;
 
 @Service
 public class AuthServiceImpl implements AuthService {
@@ -31,15 +31,15 @@ public class AuthServiceImpl implements AuthService {
   }
 
   @Override
-  public boolean register(RegisterReq registerReq, Role role) {
-    if (manager.userExists(registerReq.getUsername())) {
+  public boolean register(RegisterDto registerDto, Role role) {
+    if (manager.userExists(registerDto.getUsername())) {
       return false;
     }
     manager.createUser(
         User.builder()
             .passwordEncoder(this.encoder::encode)
-            .password(registerReq.getPassword())
-            .username(registerReq.getUsername())
+            .password(registerDto.getPassword())
+            .username(registerDto.getUsername())
             .roles(role.name())
             .build());
     return true;
