--- src/main/java/ru/skypro/homework/HomeworkApplication.java
@@ -1,11 +1,0 @@
-package ru.skypro.homework;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HomeworkApplication {
-  public static void main(String[] args) {
-    SpringApplication.run(HomeworkApplication.class, args);
-  }
-}
