--- src/main/java/ru/skypro/homework/dto/LoginReq.java
@@ -1,10 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-@Data
-public class LoginReq {
-    private String password;
-    private String username;
-
-}
