--- src/main/java/ru/skypro/homework/dto/RegisterReq.java
@@ -1,13 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-@Data
-public class RegisterReq {
-    private String username;
-    private String password;
-    private String firstName;
-    private String lastName;
-    private String phone;
-    private Role role;
-}
