--- src/test/java/ru/skypro/homework/HomeworkApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.skypro.homework;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HomeworkApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
